Mostly refactoring with some additions.
Refactored the polygons to not be full 1280x720 canvas pieces but just skewed divs instead.  Added the thin white bars on the top/bottom but suspect I'll need to use clip-path for that or maybe a small PNG image.  Can't figure out how to do the shimmer across the image and the large blue glowing effects in the background.

Added video reference file for outside consultation.